document-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
prometheus-fastapi-instrumentator==6.1.0
aiosqlite==0.19.0
aio-pika==9.3.1
openai==1.6.1
scikit-learn==1.3.2
redis==5.0.1
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
aiofiles==23.2.1
python-docx==1.1.0
langdetect==1.0.9
celery[redis]==5.3.6
